fix: add action bar to home screen
