+++ apps/mobile/src/locales/en/messages.po
@@ -26,13 +26,17 @@ msgstr ""
 msgid "{children}"
 msgstr ""
 
-#: src/app/wallet/(tabs)/action-bar.tsx:38
+#: src/components/action-bar/container.tsx:138
 msgid "{label}"
 msgstr ""
 
 #: src/app/waiting-list/waiting-list.tsx:324
 msgid "<0><1>Bitcoin for</1></0><2><3>the rest of us</3></2>"
 msgstr ""
+
+#: src/components/action-bar/container.tsx:156
+msgid "Add Wallet"
+msgstr "Add Wallet"
 
 #: src/app/wallet/developer-console/index.tsx:63
 msgid "Create wallet"
@@ -74,6 +78,10 @@ msgstr ""
 msgid "Page"
 msgstr ""
 
+#: src/components/action-bar/container.tsx:174
+msgid "Receive"
+msgstr "Receive"
+
 #: src/app/wallet/developer-console/index.tsx:91
 msgid "register for Push notifications"
 msgstr ""
@@ -82,6 +90,10 @@ msgstr ""
 msgid "schedule dummy notifications in 3s"
 msgstr ""
 
+#: src/components/action-bar/container.tsx:167
+msgid "Send"
+msgstr "Send"
+
 #: src/app/wallet/secure-your-wallet.tsx:68
 msgid "Skip for now"
 msgstr ""
@@ -89,6 +101,10 @@ msgstr ""
 #: src/app/waiting-list/waiting-list.tsx:299
 msgid "Submit"
 msgstr ""
+
+#: src/components/action-bar/container.tsx:181
+msgid "Swap"
+msgstr "Swap"
 
 #: src/app/wallet/developer-console/index.tsx:71
 msgid "Trigger API presets"
